requirements.txt
cassandra-driver==3.14.0
certifi==2018.4.16
chardet==3.0.4
cymem==1.31.2
cytoolz==0.9.0.1
dill==0.2.8.2

idna==2.7
langdetect==1.0.7
msgpack==0.5.6
msgpack-numpy==0.4.3.1
murmurhash==0.28.0
numpy==1.15.0
pandas==0.23.4
plac==0.9.6
preshed==1.0.1
python-dateutil==2.7.3
pytz==2018.5
regex==2017.4.5
requests==2.19.1
schedule==0.5.0
scikit-learn==0.19.2
scipy==1.1.0
six==1.11.0
sklearn-pandas==1.6.0
spacy==2.0.12
thinc==6.10.3
toolz==0.9.0
tqdm==4.24.0
ujson==1.35
urllib3==1.23
wrapt==1.10.11
